make sure adj_matrixt.txt is saved astype int to save storage

+++ Modules/simulation.py
@@ -96,7 +96,7 @@ class Simulation:
         
         if parameters.save_adj_matrix:
             adj_matrix = cell.compute_directed_adjacency_matrix()
-            np.savetxt(os.path.join(parameters.path, "adj_matrix.txt"), adj_matrix)
+            np.savetxt(os.path.join(parameters.path, "adj_matrix.txt"), adj_matrix.astype(int))
 
         # Classify segments by morphology, save coordinates
         segments, seg_data = cell.get_segments(["all"]) # (segments is returned here to preserve NEURON references)
